Fix up plugin loading that has to load types for Bifs.
If a plugin had a bif file with event definitions that used script
land defined types, things weren't loading in the correct order.

Now, cmake is building a script with loads all of the __preload__.zeek
scripts and loading it in the init-frameworks-and-bifs.zeek script so
that it loads before all of the bif definitions load.